Text boxes now scroll with new lines.
i.e. when it goes to a new line, if there's no space, the previous stuff
will scroll off the top, rather than simply falling out of the text box.

Also, the stuff that was already there elegantly scrolls off the top of
the box, rather than being cleared for each new line.