Second implementation of party-member following.
This one is a lot better. The follower still doesn't mimic your actions
_exactly_, but it's a good enough approximation for regular walking, and
they cover all the same ground that you cover.

The times when it doesn't mimic your actions are when you're moving back
and forth very fast, and when you walk into a wall and keep walking. In
these scenarios, it tends to just stand still and look at you. I think
this is fine -- it just looks like they consider you kind of stupid for
doing whatever you're doing, and refuse to take part.